LLM-PHASE1/chat_app/requirements.txt
streamlit
fastapi
openai==0.28
uvicorn
python-dotenv 